+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.16.7
+  * Allow message extraction to find messages in class field declarations
+    and top-level declarations.
+  * Fix incorrect name and parameters propagation during extraction phase.
+
 ## 0.16.6
   * More uppercase constant removal.